add endian.h
