+++ cpp/src/arrow/util/bit_block_counter.cc
@@ -22,7 +22,10 @@
 #include <type_traits>
 
 #include "arrow/buffer.h"
+#include "arrow/util/bit_util.h"
 #include "arrow/util/bitmap_ops.h"
+#include "arrow/util/endian.h"
+#include "arrow/util/ubsan.h"
 
 namespace arrow {
 namespace internal {
